move logic to index
