+++ wasp-app-runner/src/index.ts
@@ -1,10 +1,10 @@
+import { readdir } from "fs/promises";
 import { type Mode, parseArgs, PathToApp, WaspCliCmd } from "./args.js";
 import { startAppInBuildMode } from "./build/index.js";
 import { checkDependencies } from "./dependencies.js";
 import { startAppInDevMode } from "./dev/index.js";
 import { createLogger } from "./logging.js";
 import { waspInfo, waspTsSetup } from "./waspCli.js";
-import { isWaspTypescriptConfigProject } from "./waspTsConfig.js";
 
 const logger = createLogger("main");
 
@@ -75,5 +75,17 @@ async function runWaspApp({
 
     default:
       mode satisfies never;
+  }
+}
+
+async function isWaspTypescriptConfigProject(pathToApp: PathToApp) {
+  const logger = createLogger("wasp-ts-config");
+
+  try {
+    const files = await readdir(pathToApp);
+    return files.some((file) => file.endsWith(".wasp.ts"));
+  } catch (error) {
+    logger.error(`Failed to read directory ${pathToApp}: ${error}`);
+    process.exit(1);
   }
 }
